# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 见微 Chinese social context training API
servers:
  - url: /api
    description: Base API path
security:
  - bearerAuth: []
tags:
  - name: health
  - name: config
  - name: events
  - name: predictions
  - name: runs
  - name: export
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      security: []
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /config:
    get:
      operationId: getConfig
      tags: [config]
      summary: Get public application configuration
      security: []
      responses:
        "200":
          description: Application configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppConfig"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /summary:
    get:
      operationId: getSummary
      tags: [events]
      summary: Get the current user's event summary
      responses:
        "200":
          description: Event summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"
        "401":
          $ref: "#/components/responses/ErrorResponse"
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List the current user's events
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        "401":
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
  /events/{id}:
    parameters:
      - $ref: "#/components/parameters/EventId"
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      responses:
        "204":
          description: Event deleted
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /events/{id}/prediction:
    parameters:
      - $ref: "#/components/parameters/EventId"
    post:
      operationId: submitPrediction
      tags: [events]
      summary: Save a prediction selected by the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredictionInput"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
  /events/{id}/outcome:
    parameters:
      - $ref: "#/components/parameters/EventId"
    post:
      operationId: recordOutcome
      tags: [events]
      summary: Record a prediction outcome
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutcomeInput"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
  /events/{id}/reflection:
    parameters:
      - $ref: "#/components/parameters/EventId"
    post:
      operationId: saveReflection
      tags: [events]
      summary: Save a reflection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReflectionInput"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
  /events/{id}/analyze:
    parameters:
      - $ref: "#/components/parameters/EventId"
    post:
      operationId: analyzeEvent
      tags: [events]
      summary: Generate AI candidates for explicit user review and saving
      description: AI output is advisory only and is not persisted as an event decision.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisInput"
      responses:
        "200":
          description: Analysis candidates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
  /predictions:
    get:
      operationId: listPredictions
      tags: [predictions]
      summary: List events that have predictions
      responses:
        "200":
          description: Events with predictions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        "401":
          $ref: "#/components/responses/ErrorResponse"
  /runs:
    get:
      operationId: listRuns
      tags: [runs]
      summary: List AI analysis runs
      responses:
        "200":
          description: Analysis runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"
        "401":
          $ref: "#/components/responses/ErrorResponse"
  /export:
    get:
      operationId: exportData
      tags: [export]
      summary: Export the current user's data
      responses:
        "200":
          description: Exported data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportData"
        "401":
          $ref: "#/components/responses/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth access token
  parameters:
    EventId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
    AppConfig:
      type: object
      required: [supabaseUrl, publishableKey, aiEnabled]
      properties:
        supabaseUrl:
          type: string
          format: uri
        publishableKey:
          type: string
        aiEnabled:
          type: boolean
    Summary:
      type: object
      required:
        - totalEvents
        - drafts
        - pendingPredictions
        - duePredictions
        - recordedOutcomes
        - unknownOutcomes
        - confirmedReflections
        - completedLoops
      properties:
        totalEvents:
          type: integer
          minimum: 0
        drafts:
          type: integer
          minimum: 0
        pendingPredictions:
          type: integer
          minimum: 0
        duePredictions:
          type: integer
          minimum: 0
        recordedOutcomes:
          type: integer
          minimum: 0
          description: Predictions that have any recorded outcome, including unknown.
        unknownOutcomes:
          type: integer
          minimum: 0
          description: Recorded outcomes that cannot be judged. Excluded from completedLoops.
        confirmedReflections:
          type: integer
          minimum: 0
          description: Reflections confirmed by the user, including those after an unknown outcome.
        completedLoops:
          type: integer
          minimum: 0
          description: Locked prediction, decidable outcome, and a user-confirmed reflection. Unknown outcomes are excluded.
    EventInput:
      type: object
      required: [title, rawText, category]
      properties:
        title:
          type: string
        rawText:
          type: string
        category:
          $ref: "#/components/schemas/EventCategory"
    EventUpdate:
      type: object
      required: [version]
      properties:
        version:
          type: integer
          minimum: 1
          description: Current entity version for optimistic concurrency.
        title:
          type: string
        rawText:
          type: string
        facts:
          type: array
          items:
            type: string
        inferences:
          type: array
          items:
            type: string
        unknowns:
          type: array
          items:
            type: string
        questions:
          type: array
          maxItems: 3
          items:
            type: string
        context:
          type: string
        userHypothesis:
          type: string
        hypotheses:
          type: array
          items:
            $ref: "#/components/schemas/Hypothesis"
    Event:
      type: object
      required:
        - id
        - title
        - rawText
        - category
        - status
        - version
        - createdAt
        - updatedAt
        - facts
        - inferences
        - unknowns
        - questions
        - context
        - userHypothesis
        - hypotheses
        - prediction
        - reflection
        - reflectionCorrections
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        rawText:
          type: string
        category:
          $ref: "#/components/schemas/EventCategory"
        status:
          $ref: "#/components/schemas/EventStatus"
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        facts:
          type: array
          items:
            type: string
        inferences:
          type: array
          items:
            type: string
        unknowns:
          type: array
          items:
            type: string
        questions:
          type: array
          items:
            type: string
        context:
          type: string
        userHypothesis:
          type: string
        hypotheses:
          type: array
          items:
            $ref: "#/components/schemas/Hypothesis"
        prediction:
          allOf:
            - $ref: "#/components/schemas/Prediction"
          nullable: true
        reflection:
          allOf:
            - $ref: "#/components/schemas/Reflection"
          nullable: true
        reflectionCorrections:
          type: array
          items:
            $ref: "#/components/schemas/ReflectionCorrection"
    EventCategory:
      type: string
      enum: [work, relationship, family, other]
    EventStatus:
      type: string
      enum: [draft, analyzing, predicted, completed]
    Hypothesis:
      type: object
      required: [text, evidence, missing]
      properties:
        text:
          type: string
        evidence:
          type: string
        missing:
          type: string
    Prediction:
      type: object
      required: [id, text, probability, criteria, dueAt, createdAt, outcome]
      properties:
        id:
          type: string
          format: uuid
        text:
          type: string
        probability:
          type: integer
          minimum: 0
          maximum: 100
        criteria:
          type: string
        dueAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        outcome:
          allOf:
            - $ref: "#/components/schemas/Outcome"
          nullable: true
    Outcome:
      type: object
      required: [result, notes, intervention, createdAt]
      properties:
        result:
          type: string
          enum: [occurred, not_occurred, unknown]
        notes:
          type: string
        intervention:
          type: string
        createdAt:
          type: string
          format: date-time
    Reflection:
      type: object
      required: [lesson, wellFounded, overreach, missing, createdAt]
      properties:
        lesson:
          type: string
        wellFounded:
          type: string
        overreach:
          type: string
        missing:
          type: string
        createdAt:
          type: string
          format: date-time
    ReflectionCorrection:
      type: object
      required: [lesson, wellFounded, overreach, missing, source, createdAt]
      properties:
        lesson:
          type: string
        wellFounded:
          type: string
        overreach:
          type: string
        missing:
          type: string
        source:
          type: string
          enum: [user]
        createdAt:
          type: string
          format: date-time
    PredictionInput:
      type: object
      required: [version, requestId, text, probability, criteria, dueAt]
      properties:
        version:
          type: integer
          minimum: 1
        requestId:
          type: string
          format: uuid
        text:
          type: string
        probability:
          type: integer
          minimum: 0
          maximum: 100
        criteria:
          type: string
        dueAt:
          type: string
          format: date-time
    OutcomeInput:
      type: object
      required: [version, requestId, result, notes, intervention]
      properties:
        version:
          type: integer
          minimum: 1
        requestId:
          type: string
          format: uuid
        result:
          type: string
          enum: [occurred, not_occurred, unknown]
        notes:
          type: string
        intervention:
          type: string
    ReflectionInput:
      type: object
      required: [version, lesson, wellFounded, overreach, missing]
      properties:
        version:
          type: integer
          minimum: 1
        lesson:
          type: string
        wellFounded:
          type: string
        overreach:
          type: string
        missing:
          type: string
    AnalysisInput:
      type: object
      required: [version, requestId, stage, consent]
      properties:
        version:
          type: integer
          minimum: 1
        requestId:
          type: string
          format: uuid
        stage:
          $ref: "#/components/schemas/AnalysisStage"
        consent:
          type: boolean
        apiKey:
          type: string
          description: Optional per-request model key. Used only for that call and not stored.
        baseUrl:
          type: string
          format: uri
        model:
          type: string
    AnalysisResult:
      type: object
      required:
        - runId
        - stage
        - facts
        - inferences
        - unknowns
        - questions
        - hypotheses
        - predictionSuggestion
        - criteriaSuggestion
        - wellFounded
        - overreach
        - missing
        - lesson
      properties:
        runId:
          type: string
          format: uuid
        stage:
          $ref: "#/components/schemas/AnalysisStage"
        facts:
          type: array
          items:
            type: string
        inferences:
          type: array
          items:
            type: string
        unknowns:
          type: array
          items:
            type: string
        questions:
          type: array
          items:
            type: string
        hypotheses:
          type: array
          items:
            $ref: "#/components/schemas/Hypothesis"
        predictionSuggestion:
          type: string
        criteriaSuggestion:
          type: string
        wellFounded:
          type: string
        overreach:
          type: string
        missing:
          type: string
        lesson:
          type: string
    AnalysisStage:
      type: string
      enum: [organize, hypotheses, prediction, reflection]
    Run:
      type: object
      required:
        - id
        - eventId
        - stage
        - status
        - createdAt
        - finishedAt
        - errorCode
        - model
        - promptVersion
      properties:
        id:
          type: string
          format: uuid
        eventId:
          type: string
          format: uuid
        stage:
          $ref: "#/components/schemas/AnalysisStage"
        status:
          type: string
          enum: [running, succeeded, failed, expired]
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        errorCode:
          type: string
          nullable: true
        model:
          type: string
        promptVersion:
          type: string
    ExportData:
      type: object
      required: [exportedAt, events, runs]
      properties:
        exportedAt:
          type: string
          format: date-time
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        runs:
          type: array
          items:
            $ref: "#/components/schemas/Run"