# lib/api-zod/src/generated/types/analysisStage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * 见微 Chinese social context training API
 * OpenAPI spec version: 0.1.0
 */

export type AnalysisStage = typeof AnalysisStage[keyof typeof AnalysisStage];


export const AnalysisStage = {
  organize: 'organize',
  hypotheses: 'hypotheses',
  prediction: 'prediction',
  reflection: 'reflection',
} as const;