# lib/api-zod/src/generated/types/error.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * 见微 Chinese social context training API
 * OpenAPI spec version: 0.1.0
 */

export interface Error {
  error: string;
  code?: string;
}