# lib/api-zod/src/generated/types/outcome.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * 见微 Chinese social context training API
 * OpenAPI spec version: 0.1.0
 */
import type { OutcomeResult } from './outcomeResult';

export interface Outcome {
  result: OutcomeResult;
  notes: string;
  intervention: string;
  createdAt: Date;
}