# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * 见微 Chinese social context training API
 * OpenAPI spec version: 0.1.0
 */
import type { Outcome } from './outcome';

export interface Prediction {
  id: string;
  text: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  probability: number;
  criteria: string;
  dueAt: Date;
  createdAt: Date;
  outcome: Outcome | null;
}