# lib/api-zod/src/generated/types/predictionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * 见微 Chinese social context training API
 * OpenAPI spec version: 0.1.0
 */

export interface PredictionInput {
  /** @minimum 1 */
  version: number;
  requestId: string;
  text: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  probability: number;
  criteria: string;
  dueAt: Date;
}