# lib/api-zod/src/generated/types/reflection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * 见微 Chinese social context training API
 * OpenAPI spec version: 0.1.0
 */

export interface Reflection {
  lesson: string;
  wellFounded: string;
  overreach: string;
  missing: string;
  createdAt: Date;
}