# lib/api-zod/src/generated/types/reflectionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * 见微 Chinese social context training API
 * OpenAPI spec version: 0.1.0
 */

export interface ReflectionInput {
  /** @minimum 1 */
  version: number;
  lesson: string;
  wellFounded: string;
  overreach: string;
  missing: string;
}